third_party/blink/web_tests/fast/dom/shadow/contains-with-shadow-dom-expected.txt

CONSOLE WARNING: line 87: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Tests for node's contains().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.contains(getNodeInComposedTree("shadowHostA/")) is false
PASS getNodeInComposedTree("shadowHostA/").contains(document) is false
PASS document.contains(getNodeInComposedTree("shadowHostA/divA")) is false
PASS getNodeInComposedTree("shadowHostA/divA").contains(document) is false
PASS getNodeInComposedTree("shadowHostA/").contains(getNodeInComposedTree("shadowHostA/divA")) is true
PASS getNodeInComposedTree("shadowHostA/divA").contains(getNodeInComposedTree("shadowHostA/")) is false
PASS getNodeInComposedTree("shadowHostA/").contains(getNodeInComposedTree("shadowHostA/shadowHostB/")) is false
PASS getNodeInComposedTree("shadowHostA/shadowHostB/").contains(getNodeInComposedTree("shadowHostA/")) is false
PASS getNodeInComposedTree("shadowHostA/divA").contains(getNodeInComposedTree("shadowHostA/shadowHostB/divB")) is false
PASS getNodeInComposedTree("shadowHostA/shadowHostB/divB").contains(getNodeInComposedTree("shadowHostA/divA")) is false
PASS getNodeInComposedTree("shadowHostA/shadowHostB/").contains(getNodeInComposedTree("shadowHostA/shadowHostC/")) is false
PASS getNodeInComposedTree("shadowHostA/shadowHostC/").contains(getNodeInComposedTree("shadowHostA/shadowHostB/")) is false
PASS getNodeInComposedTree("shadowHostA/shadowHostB/divB").contains(getNodeInComposedTree("shadowHostA/shadowHostC/divC")) is false
PASS getNodeInComposedTree("shadowHostA/shadowHostC/divC").contains(getNodeInComposedTree("shadowHostA/shadowHostB/divB")) is false
PASS getNodeInComposedTree("shadowHostA/shadowHostB").contains(getNodeInComposedTree("shadowHostA/shadowHostB")) is true
PASS getNodeInComposedTree("shadowHostA/shadowHostB/").contains(getNodeInComposedTree("shadowHostA/shadowHostB/")) is true
PASS successfullyParsed is true

TEST COMPLETE
